requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.6
werkzeug==2.2.3
flask-cors==3.0.10
